feat: glibc extension
The extension is pretty much copied from the package that is in the
nvidia extensions, with one major modification: a symbolic link to
ldconfig is installed at /sbin/ldconfig. This change allows the nvidia
gpu-operator to work without modification. This change does require a
patch to the extension validation logic.

Signed-off-by: Jean-Francois Roy <jf@devklog.net>